Address sign-compare warning in game_log.cc

+++ src/core/server/connector/game_log.cc
@@ -32,7 +32,7 @@ void ExecutionData::SerializeToString(string* output) const {
   ss << "{";
   ss << "'keys':";
   ss << "[";
-  for (int i = 0; i < keys.size(); i++) {
+  for (size_t i = 0; i < keys.size(); i++) {
     if (i > 0) {
       ss << ",";
     }
@@ -83,7 +83,7 @@ void PlayerLog::SerializeToString(string* output) const {
   ss << player_id;
   ss << ",";
   ss << "'received_data':[";
-  for (int i = 0; i < received_data.size(); i++) {
+  for (size_t i = 0; i < received_data.size(); i++) {
     if (i > 0) {
       ss << ",";
     }
@@ -109,7 +109,7 @@ void GameLog::SerializeToString(string* output) const {
   ss << "'GameResult':" << result << ",";
   ss << "'error_log':'" << error_log << "',";
   ss << "'log':[";
-  for (int i = 0; i < log.size(); i++) {
+  for (size_t i = 0; i < log.size(); i++) {
     if (i > 0) {
       ss << ",";
     }
